pyproject.toml
[project]
name = "trustpartners-assistant"
version = "0.1.0"
description = "Trust Partners Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.4",
    "openai>=2.17.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
